fix: resolve port truncation by using uint16_t for ports

+++ config-proxy.h
@@ -8,9 +8,9 @@
 
 typedef struct {
   char inbound_ip[INET_ADDRSTRLEN];
-  u_int8_t inbound_port;
+  u_int16_t inbound_port;
   char outbound_host[MAX_HOST_LEN];
-  u_int8_t outbound_port;
+  u_int16_t outbound_port;
 } proxy_config_t;
 
 void parse_arguments(int argc, char *argv[], proxy_config_t *config);
